producer/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
colorama==0.4.6
Faker==40.11.0
fastapi==0.135.1
h11==0.16.0
idna==3.11
kafka-python==2.3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.42.0
